pyproject.toml
[project]
name = "snowmanpapers"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pymupdf4llm",
    "feedparser>=6.0",
    "pyyaml>=6.0",
]
